MOB-FIX : 모바일 헬스데이터 연동 버그 수정

See merge request s14-final/S14P31E105!363

+++ backend/src/main/java/com/ssafy/heygent/domain/health/dto/request/SamsungHealthRequestDto.java
@@ -18,7 +18,7 @@ public class SamsungHealthRequestDto {
     private Integer sleepDurationMinutes;
     private Double bodyFat;
     private Double skeletalMuscle;
-    private Integer bloodPressureSystolic;
-    private Integer bloodPressureDiastolic;
+    private Double bloodPressureSystolic;
+    private Double bloodPressureDiastolic;
     private Double bloodGlucose;
 }
